django/vinyl_collection/requirements.txt
asgiref==3.11.1
Django==6.0.2
psycopg2-binary==2.9.11
sqlparse==0.5.5
python-dotenv==1.0.1
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
Pillow==11.2.1
